+++ src/doomgeneric/z_zone.h
@@ -51,6 +51,7 @@ enum
         
 
 void	Z_Init (void);
+void    Z_Shutdown (void);
 void*	Z_Malloc (int size, int tag, void *ptr);
 void    Z_Free (void *ptr);
 void    Z_FreeTags (int lowtag, int hightag);